requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.4.0
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
whitenoise==6.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9  # For PostgreSQL support
dj-database-url==2.1.0  # For database URL configuration 